langgraph_react/pyproject.toml
[project]
name = "langgraph-react"
version = "0.1.0"
description = ""
authors = [
    {name = "und1n3",email = "undineunda5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.1,<0.5.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=0.3.2,<0.4.0)",
    "langchain-tavily (>=0.1.6,<0.2.0)",
    "langchain-community (>=0.3.23,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
